refactor: move utils directory to core module
