+++ privmx-endpoint/src/main/cpp/modules/Utils.cpp
@@ -9,9 +9,9 @@
 // Utils
 extern "C"
 JNIEXPORT jstring JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Utils_trim(JNIEnv *env,
-                                                                 jclass clazz,
-                                                                 jstring data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Utils_trim(JNIEnv *env,
+                                                                      jclass clazz,
+                                                                      jstring data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return nullptr;
 
@@ -32,10 +32,10 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Utils_trim(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jobject JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Utils_split(JNIEnv *env,
-                                                                  jclass clazz,
-                                                                  jstring data,
-                                                                  jstring delimiter) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Utils_split(JNIEnv *env,
+                                                                       jclass clazz,
+                                                                       jstring data,
+                                                                       jstring delimiter) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data") ||
         ctx.nullCheck(delimiter, "Delimeter"))
@@ -75,9 +75,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Utils_split(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jstring JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Utils_ltrim(JNIEnv *env,
-                                                                  jclass clazz,
-                                                                  jstring data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Utils_ltrim(JNIEnv *env,
+                                                                       jclass clazz,
+                                                                       jstring data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return nullptr;
     jstring result;
@@ -99,9 +99,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Utils_ltrim(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jstring JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Utils_rtrim(JNIEnv *env,
-                                                                  jclass clazz,
-                                                                  jstring data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Utils_rtrim(JNIEnv *env,
+                                                                       jclass clazz,
+                                                                       jstring data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return nullptr;
     jstring result;
@@ -125,9 +125,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Utils_rtrim(JNIEnv *env,
 // Hex
 extern "C"
 JNIEXPORT jstring JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Hex_encode(JNIEnv *env,
-                                                                 jclass clazz,
-                                                                 jbyteArray data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Hex_encode(JNIEnv *env,
+                                                                      jclass clazz,
+                                                                      jbyteArray data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return nullptr;
 
@@ -149,9 +149,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Hex_encode(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jbyteArray JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Hex_decode(JNIEnv *env,
-                                                                 jclass clazz,
-                                                                 jstring hex_data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Hex_decode(JNIEnv *env,
+                                                                      jclass clazz,
+                                                                      jstring hex_data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(hex_data, "Data")) return nullptr;
 
@@ -175,8 +175,8 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Hex_decode(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jboolean JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Hex_is(JNIEnv *env, jclass clazz,
-                                                             jstring data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Hex_is(JNIEnv *env, jclass clazz,
+                                                                  jstring data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return JNI_FALSE;
     jboolean result;
@@ -199,9 +199,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Hex_is(JNIEnv *env, jclass
 // Base32
 extern "C"
 JNIEXPORT jstring JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Base32_encode(JNIEnv *env,
-                                                                    jclass clazz,
-                                                                    jbyteArray data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Base32_encode(JNIEnv *env,
+                                                                         jclass clazz,
+                                                                         jbyteArray data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return nullptr;
 
@@ -223,9 +223,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Base32_encode(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jbyteArray JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Base32_decode(JNIEnv *env,
-                                                                    jclass clazz,
-                                                                    jstring base32_data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Base32_decode(JNIEnv *env,
+                                                                         jclass clazz,
+                                                                         jstring base32_data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(base32_data, "Data")) return nullptr;
 
@@ -249,9 +249,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Base32_decode(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jboolean JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Base32_is(JNIEnv *env,
-                                                                jclass clazz,
-                                                                jstring data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Base32_is(JNIEnv *env,
+                                                                     jclass clazz,
+                                                                     jstring data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return JNI_FALSE;
     jboolean result;
@@ -274,9 +274,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Base32_is(JNIEnv *env,
 // Base64
 extern "C"
 JNIEXPORT jstring JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Base64_encode(JNIEnv *env,
-                                                                    jclass clazz,
-                                                                    jbyteArray data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Base64_encode(JNIEnv *env,
+                                                                         jclass clazz,
+                                                                         jbyteArray data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return nullptr;
 
@@ -298,9 +298,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Base64_encode(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jbyteArray JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Base64_decode(JNIEnv *env,
-                                                                    jclass clazz,
-                                                                    jstring base64_data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Base64_decode(JNIEnv *env,
+                                                                         jclass clazz,
+                                                                         jstring base64_data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(base64_data, "Data")) return nullptr;
 
@@ -324,9 +324,9 @@ Java_com_simplito_java_privmx_1endpoint_modules_utils_Base64_decode(JNIEnv *env,
 
 extern "C"
 JNIEXPORT jboolean JNICALL
-Java_com_simplito_java_privmx_1endpoint_modules_utils_Base64_is(JNIEnv *env,
-                                                                jclass clazz,
-                                                                jstring data) {
+Java_com_simplito_java_privmx_1endpoint_modules_core_utils_Base64_is(JNIEnv *env,
+                                                                     jclass clazz,
+                                                                     jstring data) {
     JniContextUtils ctx(env);
     if (ctx.nullCheck(data, "Data")) return JNI_FALSE;
     jboolean result;
