+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/modules/core/utils/Base32.java
@@ -1,28 +1,28 @@
-package com.simplito.java.privmx_endpoint.modules.utils;
+package com.simplito.java.privmx_endpoint.modules.core.utils;
 
-public class Base64 {
+public class Base32 {
     static {
         System.loadLibrary("privmx-endpoint-java");
     }
 
     /**
-     * Encodes byte array to string in Base64 format.
+     * Encodes byte array to string in Base32 format.
      *
      * @param data as byte array to encode
-     * @return string in Base64 format
+     * @return string in Base32 format
      */
     public static native String encode(byte[] data);
 
     /**
-     * Decodes string in Base64 to byte array.
+     * Decodes string in Base32 to byte array.
      *
-     * @param base64_data string to decode
+     * @param base32_data string to decode
      * @return byte array with decoded data
      */
-    public static native byte[] decode(String base64_data);
+    public static native byte[] decode(String base32_data);
 
     /**
-     * Checks if given string is in Base64 format.
+     * Checks if given string is in Base32 format.
      *
      * @param data string to check
      * @return data check result
