+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/modules/core/utils/Base64.java
@@ -1,31 +1,28 @@
-package com.simplito.java.privmx_endpoint.modules.utils;
+package com.simplito.java.privmx_endpoint.modules.core.utils;
 
-public class Hex {
+public class Base64 {
     static {
         System.loadLibrary("privmx-endpoint-java");
     }
 
-    public Hex() {
-    }
-
     /**
-     * Encodes byte array to a string in Hex format.
+     * Encodes byte array to string in Base64 format.
      *
      * @param data as byte array to encode
-     * @return string in Hex format
+     * @return string in Base64 format
      */
     public static native String encode(byte[] data);
 
     /**
-     * Decodes string in Hex to byte array.
+     * Decodes string in Base64 to byte array.
      *
-     * @param hex_data string to decode
+     * @param base64_data string to decode
      * @return byte array with decoded data
      */
-    public static native byte[] decode(String hex_data);
+    public static native byte[] decode(String base64_data);
 
     /**
-     * Checks if given string is in Hex format.
+     * Checks if given string is in Base64 format.
      *
      * @param data string to check
      * @return data check result
