+++ privmx-endpoint/src/main/java/com/simplito/java/privmx_endpoint/modules/core/utils/Utils.java
@@ -1,4 +1,4 @@
-package com.simplito.java.privmx_endpoint.modules.utils;
+package com.simplito.java.privmx_endpoint.modules.core.utils;
 
 import java.util.List;
